+++ src/test/java/com/warehouse/service/StockReserveServiceImplTest.java
@@ -99,7 +99,7 @@ public class StockReserveServiceImplTest {
 
         when(stockRepository.findByItemIdForUpdate(item.getId())).thenReturn(Optional.of(stock));
 
-        when(stockReserveRepository.findSumReserveByStockAndStatus(stock, ReservationStatus.ACTIVE)).thenReturn(0L);
+        when(stockReserveRepository.findSumReserveByStockAndStatus(stock, ReservationStatus.ACTIVE)).thenReturn(0);
 
         when(userRepository.getReferenceById(ctx.userId())).thenReturn(user);
 
@@ -155,7 +155,7 @@ public class StockReserveServiceImplTest {
         when(stockRepository.findByItemIdForUpdate(item.getId())).thenReturn(Optional.of(stock));
 
         when(stockReserveRepository.findSumReserveByStockAndStatus(stock, ReservationStatus.ACTIVE)).thenReturn(
-                (long) oldReservation.getQuantity());
+                oldReservation.getQuantity());
 
         assertThrows(InsufficientStockException.class, () -> service.reserve(item.getId(), request, ctx));
 
@@ -176,7 +176,7 @@ public class StockReserveServiceImplTest {
         when(stockRepository.findByItemIdForUpdate(item.getId())).thenReturn(Optional.of(stock));
 
         when(stockReserveRepository.findSumReserveByStockAndStatus(stock, ReservationStatus.ACTIVE)).thenReturn(
-                (long) reservation.getQuantity() + reservation1.getQuantity());
+                reservation.getQuantity() + reservation1.getQuantity());
 
         assertThrows(InsufficientStockException.class, () -> service.reserve(item.getId(), request, ctx));
     }
